Module mini_fs::caseless

This module contains a caseless filesystem.

A caseless filesystem wraps an inner filesystem and treats paths as case-insensitive, regardless of the case of the inner filesystem.

Case-insensitive paths are refered to as caseless paths. A caseless path that matches the real path of a file always opens that file. Otherwise a caseless path will open the first path of the inner filesystem that matches the caseless path.

Internally paths are strings of type OsString, which can contain invalid utf8. There is no safe way to make case-insensitive comparisons when invalid utf8 is present. To minimize the effect of this restriction, the path components are compared individually. Path components with valid utf8 are compared in a case-insensitive way. Path components with invalid utf8 are compared raw (case-sensitive).

Structs

CaselessFs
Caseless filesystem wrapping an inner filesystem.