Struct mini_fs::caseless::CaselessFs

pub struct CaselessFs<S> { /* fields omitted */ }

Caseless filesystem wrapping an inner filesystem.

Methods

impl<S: Store> CaselessFs<S>

pub fn new(inner: S) -> Self

Creates a new caseless filesystem with the provided inner filesystem. It treats paths as case-insensitive, regardless of the case of the inner filesystem.

pub fn into_inner(self) -> S

Moves the inner filesystem out of the caseless filesystem. Inspired by std::io::Cursor.

pub fn get_ref(&self) -> &S

Gets a reference to the inner filesystem. Inspired by std::io::Cursor.

pub fn get_mut(&mut self) -> &mut S

Gets a mutable reference to the inner filesystem. Inspired by std::io::Cursor.

pub fn find<P: AsRef<Path>>(&self, path: P) -> Vec<PathBuf>

Finds paths that match the caseless path. Path components with valid utf8 are compared in a case-insensitive way. Path components with invalid utf8 are compared raw (case-sensitive).